Move "delete-orphans" test to "orphanage/delete"
