+++ test/functional/persistence/orphanage/delete/delete-orphans.ts
@@ -1,11 +1,11 @@
 import "reflect-metadata";
-import { Connection, Repository } from "../../../../src/index";
-import { reloadTestingDatabases, createTestingConnections, closeTestingConnections } from "../../../utils/test-utils";
+import { Connection, Repository } from "../../../../../src/index";
+import { reloadTestingDatabases, createTestingConnections, closeTestingConnections } from "../../../../utils/test-utils";
 import { expect } from "chai";
 import { Category } from "./entity/Category";
 import { Post } from "./entity/Post";
 
-describe("persistence > delete orphans", () => {
+describe("persistence > orphanage > delete", () => {
 
     // -------------------------------------------------------------------------
     // Configuration
